+++ packages/svelte/src/lib/types/general.ts
@@ -2,12 +2,11 @@ import type { Snippet } from 'svelte';
 import type { ShortcutModifierDefinition } from '@svelte-put/shortcut';
 import type {
   FitViewOptionsBase,
-  HandleType,
-  Position,
   XYPosition,
   ConnectingHandle,
   Connection,
-  OnBeforeDeleteBase
+  OnBeforeDeleteBase,
+  HandleProps as HandlePropsSystem
 } from '@xyflow/system';
 
 import type { Node } from './nodes';
@@ -24,32 +23,9 @@ export type ConnectionData = {
   connectionStatus: string | null;
 };
 
-export type HandleComponentProps = {
-  /** Type of the handle
-   * @example HandleType.Source, HandleType.Target
-   */
-  type: HandleType;
-  /** Position of the handle
-   * @example Position.TopLeft, Position.TopRight,
-   * Position.BottomLeft, Position.BottomRight
-   */
-  position?: Position;
-  /** Id of the handle
-   * @remarks optional if there is only one handle of this type
-   */
-  id?: string | null;
+export type HandleProps = HandlePropsSystem & {
   class?: string;
   style?: string;
-  /** Should you be able to connect from/to this handle */
-  isConnectable?: boolean;
-  /** Shoould you be able to connect from this handle */
-  isConnectableStart?: boolean;
-  /** Should you be able to connect to this handle */
-  isConnectableEnd?: boolean;
-  /** Function that is called when checking if connection is valid.
-   * Overrides the isValidConnection on the Flow component.
-   */
-  isValidConnection?: IsValidConnection;
   onconnect?: (connections: Connection[]) => void;
   ondisconnect?: (connections: Connection[]) => void;
   children?: Snippet;